+++ hosts/phi/configuration.nix
@@ -9,23 +9,8 @@
 
   hardware.cpu.amd.updateMicrocode = true;
 
-  # Navi10 is broken on 5.15.13+
-  boot.kernelPackages = assert (
-    # Check kernel again when 5.17 is out
-    builtins.compareVersions pkgs.linuxPackages_latest.kernel.version "5.17" == -1
-  ); pkgs.linuxPackagesFor (pkgs.linux_5_15.override {
-    argsOverride = let
-      version = "5.15.12";
-    in {
-      inherit version;
-      modDirVersion = version;
-
-      src = pkgs.fetchurl {
-        url = "mirror://kernel/linux/kernel/v5.x/linux-${version}.tar.xz";
-        sha256 = "sha256-fekZdytiZHWRUn6QTjs1g3gzgaKdgSQE9YoiJITnUaA=";
-      };
-    };
-  });
+  # Living on the edge for Navi10
+  boot.kernelPackages = pkgs.linuxPackages_latest;
 
   networking.interfaces.enp34s0.useDHCP = true;
 
@@ -43,7 +28,7 @@
   '';
 
   services.xserver.xrandrHeads = [ {
-    output = "DisplayPort-1";
+    output = "DisplayPort-2";
     primary = true;
     monitorConfig = ''
       ModeLine "2560x1440@165.08"  645.00  2560 2568 2600 2640  1440 1446 1454 1480 +hsync -vsync
@@ -51,11 +36,11 @@
     '';
   }
   {
-    output = "DisplayPort-0";
+    output = "DisplayPort-1";
     monitorConfig = ''
       ModeLine "2560x1440@165.08"  645.00  2560 2568 2600 2640  1440 1446 1454 1480 +hsync -vsync
       Option "PreferredMode" "2560x1440@165.08"
-      Option "RightOf" "DisplayPort-0"
+      Option "RightOf" "DisplayPort-2"
     '';
   } ];